Add gamma correction to tone mapper
The thing is that we are working in gamma space due to webGL, which I
think means all of our render textures are in gamma space, so whenever
we sample a texture or write to a render texture we may need to wrap our
shaders in a manual conversion from gamma to linear and visa versa.

At least that's my best guess at what's going on.