Generate and utilize indirect diffuse map
Now we first bake a base map to a render texture so we can have an
easy-to-sample unfiltered cubemap on the gpu. Then we add a proxy cube,
and apply a replacement shader, and re-render the cubemap to a CPU-side
texture. The replacement shader convolves the base cubemap into the
diffuse indirect map.

Updated the pbr shader to utilize the indirect diffuse map as part of
the ambient term.