Fix a couple pretty major issues
Since we started baking the cubemaps at runtime we didn't realize the
baking camera was right at the bottom of the teacup resulting in scuffed
bakes. Just moved baking geo to separate layer.

The other thing is the seems in the diffuse maps were apparently caused
by mip mapping of the cubemap. Unless we are doing the specular map, we
always go for lod 0