Ok we are back to baking indirect specular map at runtime
The problem was that we were doing things via copy texture which isn't
supported on webgl.
However, there is a way around it. We aim the camera in the desired
direction, shoot a standard 2D texture. Then set the active framebuffer
/ render target to our specular map with a specified face and mip. Then
we manually blit by rendering a full screen quad with a simple blit
material. The order of the vertices is important, otherwise you may get
nothing.
Also when blitting to the cubemap, you actually need to flip some of the
textures as they were rendered. I found this out the hard way.
Also also, for some reason I never fully grasped, converting local
position to normals wasn't working in the convolution shader, but going
to world space coordinates seemed to do the trick.