Little cleanup
For a bit of exploration and education decided to make the specular map
baking go two ways depending on SystemInfo.CopyTextureSupport. If we
have it, then we use CopyTexture, otherwise we use the technique
mentioned in previous commit.
This is unnecessary since the previous commit solved it for both cases,
but for my own reference I'd like to see both solutions