Start of drag to translate
Currently we only support moving along a plane not a single axis.
Also put in the little light bulb icon just for funsies

As for translation it's this whole chain of command. The arrow handle
has a draggable on it, which lets you drag a thing around and fires an
event when that happens. The gizmo itself hooks to the draggable and
fires an event that Transformable listens for which is a component that
sits on the object itself. Finally transformable does the move and fire
another event, which in the point lights case is used to update the
lighting data.