Unchild the lines from the translate gizmo
This will help us in the future when we want to draw just a line or two.
The previous parent-child setup would make it impossible to hide an
arrow and show a line at the same time, which we would want when
translating something with the g key, and then constraining that
movement to an axis or two.